+++ health-care-api/src/main/java/com/health/care/domain/Condition.java
@@ -28,20 +28,20 @@ public class Condition {
     @Field("ENCOUNTER") 
     private String encounter;
     @Field("CODE")
-    private Integer code;
+    private Long code;
     @Field("DESCRIPTION") 
     private String description;
 
     /**
      * @param code the code to set
      */
-    public void setCode(Integer code) {
+    public void setCode(Long code) {
         this.code = code;
     }
     /**
      * @return the code
      */
-    public Integer getCode() {
+    public Long getCode() {
         return code;
     }
     /**
